pyproject.toml
[project]
name = "drf-exam"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "django-filter>=25.2",
    "django-mptt>=0.18.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.1.0",
    "psycopg-binary>=3.3.2",
    "psycopg2-binary>=2.9.11",
]
